soc: arm: alif_ensemble: Image layout changes and mram partitions
Add MCUBoot compatible mram storage partitions to the mram storage
node in the device tree for rtss_he and rtss_hp. Also, replace the mram
storage node as the chosen flash node and introduce the chosen
'code-partition' node to support MCUBoot compatible application
builds. While doing this, the default builds will link images to the
MRAM. So add a config option to build images to run from the ITCM and
remove the XIP_MRAM config entry which was dependent on additional
flash nodes.

Also, for the applications that do not enforce code_partition
(especially non MCUBoot compatible application builds), add an offset
to the RTSS_HP build so that RTSS_HE and RTSS_HP builds link images to
different addresses in the MRAM.

With this, following is the default build settings(with ROM_ITCM=y).
The user can override these with config entries as needed.

ITCM build (ROM_ITCM=y): RTSS_HP and RTSS_HE images will get linked
to the start of the ITCM address.

MRAM build (ROM_ITCM=n): RTSS_HE will get linked to the start of the
MRAM storage area and RTSS_HP images will get linked to an offset
of 0x200000 from the start of the MRAM storage area.

The MCUBoot compatible MRAM builds (with ROM_ITCM=n)will link images to
the chosen "code_partition" partition address as usual.

Signed-off-by: Silesh C V <silesh@alifsemi.com>