soc: alif_ensemble: mpu: Add device memory regions
Define the device memory region for the host peripherals. While doing
this remove the custom REGION_DEVICE_ATTR helper macro that was added
to support the OSPI memory regions. This was needed earlier because
support for device memory regions was not present for the v8-m
architecture. Also, enable CPU_HAS_CUSTOM_FIXED_SOC_MPU_REGIONS for
ensemble.

Signed-off-by: Silesh C V <silesh@alifsemi.com>