dts: arm: alif: add power state for ensemble rtss
We are supporting only the SOFT_OFF state where the
subsystem will go to OFF state.

The min-residency-us parameter is set to 5min which cant be
achieved with system clock and the tick period set to 1ms.

User may force the application to invoke the SOFT_OFF
by using the pm_state_force() function.

Signed-off-by: Sudhir Sreedharan <sudhir@alifsemi.com>