soc: drivers: Add support for ARMClang linker sections
The ARMClang supports scatter files as opposed to ld scripts that are
the default supported by Zephyr linker scripts. This caused the display
framebuffers not properly get linked to desired memory sections.

Added support to ARMClang to link memory request to desired linker
sections, be it Non-secure or SRAM0/1 depending upon the availability
of the respective nodes across all board families.

Signed-off-by: Yogender Kumar Arya <yogender.kumar@alifsemi.com>